leetcode/m2368_reachable_nodes_with_restrictions/reachable_nodes_with_restricted_test.go
package reachablenodeswithrestrictions

import "testing"

func Test_reachableNodes(t *testing.T) {
	type args struct {
		n          int
		edges      [][]int
		restricted []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "1",
			args: args{
				n: 7,
				edges: [][]int{
					{0, 1},
					{1, 2},
					{3, 1},
					{4, 0},
					{5, 0},
					{5, 6},
				},
				restricted: []int{4, 5},
			},
			want: 4,
		},
		{
			name: "2",
			args: args{
				n: 7,
				edges: [][]int{
					{0, 1},
					{0, 2},
					{0, 5},
					{0, 4},
					{3, 2},
					{6, 5},
				},
				restricted: []int{4, 2, 1},
			},
			want: 3,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := reachableNodes(tt.args.n, tt.args.edges, tt.args.restricted); got != tt.want {
				t.Errorf("reachableNodes() = %v, want %v", got, tt.want)
			}
		})
	}
}
